Move form-model code up to shared lib
