+++ controllers/apply/lib/formatters.js
@@ -11,8 +11,8 @@ const sumBy = require('lodash/sumBy');
 const filesize = require('filesize');
 const mime = require('mime-types');
 
-const fromDateParts = require('../../lib/from-date-parts');
-const countWords = require('../../lib/count-words');
+const fromDateParts = require('./from-date-parts');
+const countWords = require('./count-words');
 
 function formatRadio(field) {
     return function(value) {
